web-nfc/nfc_watch.https.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Web NFC: nfc.watch tests</title>
<link rel="author" title="Intel" href="http://www.intel.com"/>
<link rel="help" href="https://w3c.github.io/web-nfc/"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/nfc_help.js"></script>

<script>

"use strict";

promise_test(async t => {
  const invalidSignals = [
    'string',
    123,
    false,
    {},
    Symbol(),
    () => {},
    self,
  ];

  for (const signal of invalidSignals) {
    await promise_rejects(
      t, new TypeError(),
      navigator.nfc.watch(noop, {signal: signal}));
  }
}, 'Test that the signal option must be an AbortSignal.');

promise_test(async t => {
  const controller = new AbortController();
  controller.abort();

  await promise_rejects(
    t, 'AbortError',
    navigator.nfc.watch(noop, {signal: controller.signal}));
}, 'Test that passing an already aborted signal should abort.');

promise_test(async t => {
  const controller = new AbortController();

  await promise_rejects(
    t, 'AbortError',
    new Promise(resolve => {
      navigator.nfc.watch(noop, {signal: controller.signal});
      controller.abort();
      resolve();
    }));
}, 'Test that an aborted request results in AbortError.');

promise_test(t => {
  const controller = new AbortController();

  assert_false(controller.signal.aborted);
  return navigator.nfc.watch(noop, {signal: controller.signal});
}, 'Test that no exception occurs if signal is not aborted.');

promise_test(t => {
  return navigator.nfc.watch(noop);
}, "Test that nfc watch succeeds if NFC HW is enable.");

promise_test(t => {
  return promise_rejects(t, 'SyntaxError', navigator.nfc.watch(noop, {url:"www.a.com"}));
}, 'Test that nfc.watch fails if NFCWatchOptions.url is missing components.');

promise_test(t => {
  return promise_rejects(t, 'SyntaxError', navigator.nfc.watch(noop, {url:"invalid"}));
}, 'Test that nfc.watch fails if NFCWatchOptions.url is invalid.');

promise_test(t => {
  return promise_rejects(t, 'SyntaxError', navigator.nfc.watch(noop, {url:"http://a.com"}));
}, 'Test that nfc.watch fails if NFCWatchOptions.url has wrong protocol.');

promise_test(t => {
  return navigator.nfc.watch(noop, {url:"https://a.com"});
}, 'Test that nfc.watch succeeds if NFCWatchOptions.url is valid URL.');

promise_test(t => {
  return navigator.nfc.watch(noop, {url:"https://a.com/*"});
}, 'Test that nfc.watch succeeds if NFCWatchOptions.url is valid URL with "*"' +
   ' wildcard character in path.');

promise_test(t => {
  return navigator.nfc.watch(noop, {url:"https://a.com/*/bar"});
}, 'Test that nfc.watch succeeds if NFCWatchOptions.url is valid URL with "*"' +
   ' wildcard character in the beginning of path component followed by' +
   ' subpath.');

promise_test(t => {
  return navigator.nfc.watch(noop, {url:""});
}, 'Test that nfc.watch succeeds if NFCWatchOptions.url is empty.');

</script>
